Minor change for compotatiblity with dendrite analysis

+++ CNMF-e/ca_source_extraction/@Sources2D/post_process_spatial.m
@@ -30,5 +30,5 @@ parfor m=1:K
     
     A_new{m} = ai(:);
 end
-A_ = sparse(cell2mat(A_new));
+A_ = sparse(double(cell2mat(A_new)));
 end
